+++ empire/server/stagers/multi/launcher.py
@@ -83,6 +83,8 @@ class Stager:
                 "Description": "Bypasses as a space separated list to be prepended to the launcher",
                 "Required": False,
                 "Value": "",
+                "BypassLanguage": "powershell",
+                "BypassLanguageOverrides": {"python": "python"},
             },
         }
 
@@ -114,6 +116,7 @@ class Stager:
                 obfuscation_command=obfuscate_command,
                 encode=encode,
                 listener_name=listener_name,
+                bypasses=self.options["Bypasses"]["Value"],
             )
 
         elif language == "go":
@@ -123,6 +126,7 @@ class Stager:
                 encode=encode,
                 obfuscate=invoke_obfuscation,
                 obfuscation_command=obfuscate_command,
+                bypasses=self.options["Bypasses"]["Value"],
             )
         else:
             launcher = self.mainMenu.stagergenv2.generate_launcher(