+++ empire/server/stagers/windows/bunny.py
@@ -52,6 +52,7 @@ class Stager:
                 "Description": "Bypasses as a space separated list to be prepended to the launcher",
                 "Required": False,
                 "Value": "",
+                "BypassLanguage": "powershell",
             },
             "Language": {
                 "Description": "Language of the stager to generate.",
@@ -134,6 +135,7 @@ class Stager:
                 obfuscation_command=obfuscate_command,
                 encode=True,
                 listener_name=listener_name,
+                bypasses=bypasses,
             )
 
         if not launcher: