+++ empire/server/stagers/windows/cmd_exec.py
@@ -72,6 +72,7 @@ class Stager:
                 "Description": "Bypasses as a space separated list to be prepended to the launcher",
                 "Required": False,
                 "Value": "",
+                "BypassLanguage": "powershell",
             },
             "Arch": {
                 "Description": "Architecture of the .dll to generate (x64 or x86).",
@@ -114,6 +115,7 @@ class Stager:
                 obfuscation_command=obfuscate_command,
                 encode=encode,
                 listener_name=listener_name,
+                bypasses=self.options["Bypasses"]["Value"],
             )
 
         elif language == "powershell":