+++ empire/server/stagers/windows/dll.py
@@ -78,6 +78,7 @@ class Stager:
                 "Description": "Bypasses as a space separated list to be prepended to the launcher",
                 "Required": False,
                 "Value": "",
+                "BypassLanguage": "powershell",
             },
         }
 
@@ -117,6 +118,7 @@ class Stager:
                 obfuscation_command=obfuscate_command,
                 encode=True,
                 listener_name=listener_name,
+                bypasses=bypasses,
             )
 
         elif language == "powershell":