+++ empire/server/stagers/windows/macro.py
@@ -88,6 +88,7 @@ class Stager:
                 "Description": "Bypasses as a space separated list to be prepended to the launcher",
                 "Required": False,
                 "Value": "",
+                "BypassLanguage": "powershell",
             },
             "OutlookEvasion": {
                 "Description": "Include BC-Security's Outlook Sandbox evasion code",
@@ -158,6 +159,7 @@ class Stager:
                 obfuscation_command=obfuscate_command,
                 encode=encode,
                 listener_name=listener_name,
+                bypasses=bypasses,
             )
         elif language == "powershell":
             launcher = self.mainMenu.stagergenv2.generate_launcher(