+++ empire/server/stagers/windows/teensy.py
@@ -70,6 +70,7 @@ class Stager:
                 "Description": "Bypasses as a space separated list to be prepended to the launcher",
                 "Required": False,
                 "Value": "",
+                "BypassLanguage": "powershell",
             },
         }
 
@@ -96,6 +97,7 @@ class Stager:
                 obfuscation_command=obfuscate_command,
                 encode=True,
                 listener_name=listener_name,
+                bypasses=self.options["Bypasses"]["Value"],
             )
         elif language == "powershell":
             launcher = self.mainMenu.stagergenv2.generate_launcher(