+++ empire/server/stagers/windows/wmic.py
@@ -82,6 +82,7 @@ class Stager:
                 "Description": "Bypasses as a space separated list to be prepended to the launcher",
                 "Required": False,
                 "Value": "",
+                "BypassLanguage": "powershell",
             },
         }
 
@@ -111,6 +112,7 @@ class Stager:
                 obfuscation_command=obfuscate_command,
                 encode=encode,
                 listener_name=listener_name,
+                bypasses=self.options["Bypasses"]["Value"],
             )
         elif language == "powershell":
             launcher = self.mainMenu.stagergenv2.generate_launcher(